Remove <?php tag from code example; confusing.
The message says to add the following after the <?php tag, but then
includes the <?php tag in the example of what to add. We've had at least
one person think this meant having two <?php tags in their wp-config.php
which breaks things. Removing the extra <?php tag for clarity.